refactor : Car 클래스 Test 분리
