+++ src/test/java/racingcar/model/CarTest.java
@@ -1,0 +1,39 @@
+package racingcar.model;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class CarTest {
+
+    @DisplayName("자동차 이름은 공백을 제거하고 6글자 이상일 수 없다.")
+    @ParameterizedTest
+    @ValueSource(strings = {"ethaan   ", "veruasds   ", "   pasdobi\t\t\t"})
+    void inputUnderFiveLength(String carName) {
+        Assertions.assertThatThrownBy(() -> new Car(carName))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("특정 값 이상인 경우 자동차를 이동시킨다")
+    @ParameterizedTest
+    @ValueSource(ints = {4, 5, 6, 7, 8, 9})
+    void moveCar(int moveValue) {
+        Car car = new Car("a");
+
+        car.move(moveValue);
+        assertThat(car.getPosition()).isOne();
+    }
+
+    @DisplayName("특정 값 이하인 경우 자동차를 정지시킨다")
+    @ParameterizedTest
+    @ValueSource(ints = {0, 1, 2, 3})
+    void stopCar(int moveValue) {
+        Car car = new Car("a");
+
+        car.move(moveValue);
+        assertThat(car.getPosition()).isZero();
+    }
+}
